+++ src/pages/DetailPage/DetailPage.tsx
@@ -1,9 +1,10 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useEffect, useMemo, useState } from 'react';
+import { useParams } from 'react-router-dom';
 
 import { DetailInfo } from '../../components/DetailInfo';
 import { Loader } from '../../components/Loader';
 import { Character, defaultCharacter } from '../../types/types';
+import { GetRef } from '../../utils/GetRef';
 import { getCharacter } from '../../utils/api';
 import './style.css';
 
@@ -12,7 +13,6 @@ export default function DetailPage() {
   const [loading, setLoading] = useState(false);
   const params = useParams();
   const detailId = params.id ?? 0;
-  const navigate = useNavigate();
 
   useEffect(() => {
     const getData = async () => {
@@ -36,19 +36,7 @@ export default function DetailPage() {
     return <DetailInfo character={character} />;
   }, [character, loading]);
 
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleClick = (event: MouseEvent) => {
-      if (ref.current && !ref.current.contains(event.target as Node)) {
-        navigate('/', { replace: true });
-      }
-    };
-    document.addEventListener('mousedown', handleClick);
-    return () => {
-      document.removeEventListener('mousedown', handleClick);
-    };
-  }, [navigate]);
+  const ref = GetRef();
 
   return (
     <div className='cards__detail' ref={ref}>
